refactor: replace some env variables with constants

+++ backend/common/pkg/structs/mail.go
@@ -3,19 +3,19 @@ package structs
 type MailRequestVerification struct {
 	To               string
 	Otp              string
-	ExpiresInMinutes int
+	ExpiresInMinutes int16
 }
 
 type MailRequestLoginOtp struct {
 	To               string
 	Otp              string
-	ExpiresInMinutes int
+	ExpiresInMinutes int16
 }
 
 type MailRequestForgotPassword struct {
 	To               string
 	Otp              string
-	ExpiresInMinutes int
+	ExpiresInMinutes int16
 }
 
 type MailRequestResetPassword struct {
